[SYCL][NFC] Move more UR-related declarations into library
This patch also cleaned up some `#include` directives
