+++ sycl/include/sycl/detail/ur.hpp
@@ -19,7 +19,6 @@
 #include <sycl/detail/os_util.hpp>
 #include <ur_api.h>
 
-#include <memory>
 #include <type_traits>
 #include <vector>
 
@@ -103,9 +102,6 @@ __SYCL_EXPORT void contextSetExtendedDeleter(const sycl::context &constext,
                                              void *user_data);
 }
 
-class Adapter;
-using AdapterPtr = std::shared_ptr<Adapter>;
-
 // TODO: To be removed as this was only introduced for esimd which was removed.
 template <sycl::backend BE>
 __SYCL_EXPORT void *getPluginOpaqueData(void *opaquedata_arg);
@@ -122,15 +118,6 @@ int unloadOsLibrary(void *Library);
 // Function to get Address of a symbol defined in the shared
 // library, implementation is OS dependent.
 void *getOsLibraryFuncAddress(void *Library, const std::string &FunctionName);
-
-void *getURLoaderLibrary();
-
-// Performs UR one-time initialization.
-std::vector<AdapterPtr> &
-initializeUr(ur_loader_config_handle_t LoaderConfig = nullptr);
-
-// Get the adapter serving given backend.
-template <backend BE> const AdapterPtr &getAdapter();
 
 // The SYCL_UR_TRACE sets what we will trace.
 // This is a bit-mask of various things we'd want to trace.
