+++ sycl/source/platform.cpp
@@ -10,7 +10,7 @@
 #include <detail/config.hpp>
 #include <detail/global_handler.hpp>
 #include <detail/platform_impl.hpp>
-#include <sycl/detail/ur.hpp>
+#include <detail/ur.hpp>
 #include <sycl/device.hpp>
 #include <sycl/device_selector.hpp>
 #include <sycl/image.hpp>
