+++ sycl/unittests/helpers/UrMock.hpp
@@ -34,6 +34,7 @@
 #include <detail/adapter.hpp>
 #include <detail/global_handler.hpp>
 #include <detail/platform_impl.hpp>
+#include <detail/ur.hpp>
 #include <sycl/detail/common.hpp>
 #include <sycl/device.hpp>
 #include <sycl/device_selector.hpp>
